requirements.txt
category_encoders==2.5.0
feature-engine==1.2.0
gitdb==4.0.9
GitPython==3.1.27
keras==2.8.0
Keras-Preprocessing==1.1.2
kneed==0.7.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
optuna==2.10.0
pandas==1.4.1
psycopg2-binary
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
statsmodels==0.13.2
streamlit==1.8.1
tensorflow==2.8.0
urllib3==1.26.7
validators==0.18.2
xgboost==1.6.0